Moves calculator test to analyst test director.
