+++ lib/sales_engine.rb
@@ -4,8 +4,13 @@ require_relative 'invoice_repository'
 require_relative 'invoice_item_repository'
 require_relative 'customer_repository'
 require_relative 'transaction_repository'
+require_relative 'sales_engine_finders'
+require_relative 'sales_engine_creators'
 
 class SalesEngine
+
+  include SalesEngineFinders
+  include SalesEngineCreators
 
   attr_reader :merchants,
               :items,
@@ -25,92 +30,6 @@ class SalesEngine
     @invoice_items = create_invoice_item_repository(files)
     @customers = create_customer_repository(files)
     @transactions = create_transaction_repository(files)
-  end
-
-  def create_transaction_repository(files)
-    if files.include?(:transactions)
-      TransactionRepository.new(files[:transactions], self)
-    end
-  end
-
-  def create_invoice_item_repository(files)
-    if files.include?(:invoice_items)
-      InvoiceItemRepository.new(files[:invoice_items], self)
-    end
-  end
-
-  def create_customer_repository(files)
-    if files.include?(:customers)
-      CustomerRepository.new(files[:customers], self)
-    end
-  end
-
-  def create_item_repository(files)
-    if files.include?(:items)
-      ItemRepository.new(files[:items], self)
-    end
-  end
-
-  def create_merchant_repository(files)
-    if files.include?(:merchants)
-      MerchantRepository.new(files[:merchants], self)
-    end
-  end
-
-  def create_invoice_repository(files)
-    if files.include?(:invoices)
-      InvoiceRepository.new(files[:invoices], self)
-    end
-  end
-
-  def find_merchant_by_merchant_id(merchant_id)
-    merchants.find_by_id(merchant_id)
-  end
-
-  def find_invoices_by_merchant_id(merchant_id)
-    invoices.find_all_by_merchant_id(merchant_id)
-  end
-
-  def find_invoices_by_customer_id(customer_id)
-    invoices.find_all_by_customer_id(customer_id)
-  end
-
-  def find_items_by_merchant_id(merchant_id)
-    items.find_all_by_merchant_id(merchant_id)
-  end
-
-  def find_customers_by_merchant_id(merchant_id)
-    invoices = find_invoices_by_merchant_id(merchant_id)
-    customer_ids = invoices.map {|invoice| invoice.customer_id}
-    customer_ids.map {|customer_id| customers.find_by_id(customer_id)}.uniq
-  end
-
-  def find_merchants_by_customer_id(customer_id)
-    invoices = find_invoices_by_customer_id(customer_id)
-    merchant_ids = invoices.map {|invoice| invoice.merchant_id}
-    merchant_ids.map {|merchant_id| merchants.find_by_id(merchant_id)}
-  end
-
-  def find_invoice_items_by_invoice_id(invoice_id)
-    invoice_items.find_all_by_invoice_id(invoice_id)
-  end
-
-  def find_items_by_invoice_id(invoice_id)
-    invoice_items = find_invoice_items_by_invoice_id(invoice_id)
-    item_ids = invoice_items.map {|invoice_item| invoice_item.item_id}
-    item_ids.map {|item_id| items.find_by_id(item_id)}
-  end
-
-  def find_invoice_by_invoice_id(invoice_id)
-    invoices.find_by_id(invoice_id)
-  end
-
-  def find_transactions_by_invoice_id(invoice_id)
-    transactions.find_all_by_invoice_id(invoice_id)
-  end
-
-  def find_customer_by_customer_id(customer_id)
-    customers.find_by_id(customer_id)
   end
 
 end
