+++ test/analyst_tests/calculator_test.rb
@@ -1,5 +1,5 @@
 require_relative 'test_helper'
-require_relative '../lib/calculator'
+require_relative '../../lib/calculator'
 
 class CalculatorTest < Minitest::Test
 
@@ -28,6 +28,10 @@ class CalculatorTest < Minitest::Test
   def test_it_returns_0_when_calculating_standard_deviation_of_set_of_1
     set = [1]
     assert_equal 0, standard_deviation(set)
+  end
+
+  def test_percentage_returns_percent_of_two_numbers
+    assert_equal 50, percentage(1,2)
   end
 
 end
