+++ test/analyst_tests/invoice_analyst_test.rb
