+++ test/analyst_tests/merchant_analyst_test.rb
